backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Fabien Le Frapper <contact@fabienlefrapper.me>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
wagtail = "<2.11"
wagtail-grapple = "^0.8.0"
channels = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
